+++ example/nuxt.config.js
@@ -1,10 +1,14 @@
-import tailwindModule from '..'
+import tailwindModule from '../src'
 
 export default {
   buildModules: [
     tailwindModule
   ],
+
+  components: true,
+
   tailwindcss: {
+    jit: true,
     exposeConfig: true
   }
 }