+++ test/mock/tailwind.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  purge: {
-    content: [
-      'content/**/*.md'
-    ]
-  }
-}